requirements.txt
groq
vosk
requests
pyaudio
pynput
python-dotenv
PyMuPDF